// wardn/frontend/lib/api/generated/model/mCPCatalogSourceRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wardn AI API
 * OpenAPI spec version: 0.0.1
 */

export interface MCPCatalogSourceRead {
  baseUrl: string;
  createdAt: string;
  hasAuthToken: boolean;
  id: string;
  isEnabled: boolean;
  lastError: string;
  lastSuccessAt?: string | null;
  lastSyncedUpdatedSince?: string | null;
  name: string;
  organizationId: string;
  provider: string;
  syncMode: string;
  tenantId: string;
  updatedAt: string;
}